feat/templates: allow asymetric tests based on `pkg.sh.*` files
`cargo run` now compiles and picks up the templates, reads all files
existing in the directory defined by the env variable `JUICE_CONTAINERS`,
creates a per container backend list and generates the concourse yml
files based on that.
Note that the sorting has to be kept in order to keep the git diffs
minimal.